+++ src/libostree/ostree-cmd-private.h
