+++ src/libostree/ostree-repo-static-delta-core.c
@@ -25,7 +25,7 @@
 #include "ostree-core-private.h"
 #include "ostree-repo-private.h"
 #include "ostree-lzma-decompressor.h"
-#include "ostree-cmdprivate.h"
+#include "ostree-cmd-private.h"
 #include "ostree-checksum-input-stream.h"
 #include "ostree-repo-static-delta-private.h"
 #include "otutil.h"
